Fix issues with model_name variable, vLLM dash
The vLLM dashboard model_name variable will sometimes pick up CPU names
in the list of options. This is because the variable was set to just
grab model_names from anywhere in cluster. This PR restricts it to only
grab the model_name that pertains to vLLM metrics. The other issue was
that even if a model has metrics it wasn't showing up in the list. This
PR makes it so that the list of models refreshes when you select time
range so that should fix that issue.

Signed-off-by: Isaiah Stapleton <istaplet@redhat.com>